Implemented sise sync with git cli

+++ include/commands/search.h
@@ -3,7 +3,7 @@
 typedef struct searchargs {
     char **words; 
     char *searchstr;
-    int wordcount;
+    size_t wordcount;
 } searchsyms;
 
 struct threadargs {
